Hedgehog + Wolf animation events
Moved hedgehog and wolf to their own folders
Created animation events for both
