+++ Assets/Prefabs/Creatures/Mammalian/Rhino.prefab
@@ -367,8 +367,8 @@ BoxCollider:
   m_ProvidesContacts: 0
   m_Enabled: 1
   serializedVersion: 3
-  m_Size: {x: 1.5895672, y: 3.2161582, z: 2.8225784}
-  m_Center: {x: -0.038012028, y: 1.6068877, z: -0.047697008}
+  m_Size: {x: 1.5895672, y: 8, z: 2.8225784}
+  m_Center: {x: -0.038012028, y: 4, z: -0.047697008}
 --- !u!114 &260096233365818866
 MonoBehaviour:
   m_ObjectHideFlags: 0
