+++ Assets/Rendering/3D Assets/Creatures/Mammals/Hedgehog/HedgehogAttackEdit.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 17bab9ec54e7e1c4392ffd8d733dbec6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
