+++ Assets/Rendering/3D Assets/Creatures/Mammals/Wolf/AttackEdit.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1696058565057ce45b457820ae62d649
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
