--- Assets/Scripts/Creature/Major Creatures/Wolf.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f2d69942c382eb646a4e0f4dc523b0fe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
