+++ Assets/Scripts/Creature/Major Creatures/Wolf.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 66d0464642fe0dd40a1713e5af761354
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
