+++ Assets/Scripts/Creature/Major Creatures/Wolf/WolfAnimationEvents.cs
@@ -1,0 +1,23 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class WolfAnimationEvents : MonoBehaviour
+{
+    [SerializeField] private MeshCollider wolfAttackCollider;
+
+    public void EnableWolfAttackCollider()
+    {
+        wolfAttackCollider.enabled = true;
+    }
+
+    public void DisableWolfAttackCollider()
+    {
+        wolfAttackCollider.enabled = false;
+    }
+
+    public void AttackCompleted()
+    {
+        GetComponentInParent<Wolf>().attackCompleted = true;
+    }
+}
