+++ Assets/Scripts/Creature/Major Creatures/Wolf/WolfAnimationEvents.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c5efbef61c253a44db8824699f187e8a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
