--- Assets/Scripts/Creature/Minor Creatures/Hedgehog.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 94d6624bfa864a6468fc2592c612416c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
