+++ Assets/Scripts/Creature/Minor Creatures/Hedgehog.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8fd6db2f414eef04bbd338d2313eecd2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
