+++ Assets/Scripts/Creature/Minor Creatures/Hedgehog/HedgehogAnimationEvents.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1a76ea3154ef0064186d20e5945aa1f5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
